ai_inference/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
torch==2.2.0
diffusers==0.25.0
transformers==4.37.2
accelerate==0.26.1
pillow==10.2.0
grpcio==1.60.1
grpcio-tools==1.60.1
python-multipart==0.0.9
numpy==1.26.3 